refactor: seperate block rendering logic into repository, actions, and viewModel
